Fix remaining low-priority security and hardening findings
Address the final issues from the manual static security review:

- Re-stat and validate the --keepdir path after privilege drop and
  verify it is actually a directory before checking permissions.
- Check append2buffer() return values in headerchain2signingbuffer()
  so MIME-header serialization failures are propagated.
- Replace remaining strcpy/strcat uses in dict_cdb and the debug-only
  dump helpers to silence the custom 'make check' target.
- Write the in-memory BIO to disk in INT_MAX chunks to avoid the
  OpenSSL int-length truncation for very large messages.
- Use correct printf format specifiers in output_stats().

Build is clean with -Wall -Wextra, make check is silent, and all
unit tests pass.

+++ src/stats.c
@@ -63,6 +63,6 @@ void output_stats(void) {
     get_signing_time(&stats_totaltime);
     stats_count = get_signed_mails();
 
-    logmsg(LOG_NOTICE, "STATISTIK: %d/%d.%d", stats_count,
-       stats_totaltime.tv_sec,stats_totaltime.tv_usec);
+    logmsg(LOG_NOTICE, "STATISTIK: %lu/%ld.%ld", stats_count,
+       (long) stats_totaltime.tv_sec, (long) stats_totaltime.tv_usec);
 }
